@@ -1,17 +1,13 @@
 package com.graphaware.neo4j.config.model.rbac;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-@Data @AllArgsConstructor @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Role {
+public record Role(String name, boolean dropIfExists, List<Privilege> privileges) {
 
-    private String name;
-    private boolean dropIfExists;
-    private List<Privilege> privileges;
+    public Role {
+        privileges = List.copyOf(privileges);
+    }
 }
